// src/types/Phrase.d.ts
export type Phrase =
  | "Add User To Chat"
  | "Add"
  | "Admin"
  | "Are you sure?"
  | "Cancel"
  | "Chat"
  | "Create a New Chat"
  | "Delete Chat"
  | "Delete"
  | "Deleted Chat with id"
  | "Deleting chat"
  | "Email"
  | "End of Messages"
  | "Enter message in ANY language..."
  | "Get the conversation started..."
  | "Invite a friend"
  | "Let's get you started by creating your first chat!"
  | "New Chat"
  | "No messages yet"
  | "Note: They must be registered"
  | "Please wait while we delete the chat..."
  | "Send your first message in ANY language below to get started!"
  | "Send"
  | "Simply enter another users email address to invite them to this chat!"
  | "The AI will auto-detect & translate it all for you..."
  | "This will delete the chat for all users."
  | "Welcome!"